src/backend/requirements.txt
azure-storage-blob>=12.25.1,<=13.0
python-dotenv>=1.1.0
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
pymssql>=2.3.4,<2.4.0
python-multipart>=0.0.20
